+++ src/jpl_ephem/horizon/mod.rs
@@ -1,2 +1,2 @@
-pub mod horizon_reader;
+pub mod horizon_data;
 pub mod horizon_records;
