+++ src/jpl_ephem/naif/mod.rs
@@ -3,7 +3,7 @@ mod directory;
 mod ephemeris_record;
 mod jpl_ephem_header;
 mod naif_ids;
-pub mod naif_reader;
+pub mod naif_data;
 mod summary_record;
 
 pub fn print_hex_dump(data: &[u8]) {
